requirements.txt
ultralytics
easyocr
google-genai
python-dotenv
opencv-python
gTTS
requests